+++ packages/cli/test/generators.test.ts
@@ -35,7 +35,6 @@ describe('@feathersjs/cli', () => {
 
       before(async () => {
         cwd = await mkdtemp(path.join(os.tmpdir(), name + '-'))
-        console.log(cwd)
         context = await generate(
           getContext<AppGeneratorContext>(
             {
@@ -114,6 +113,14 @@ describe('@feathersjs/cli', () => {
 
         assert.ok(customServiceContext)
         assert.strictEqual(testResult, 0)
+      })
+
+      it('compiles successfully', async () => {
+        if (language === 'ts' && framework === 'koa') {
+          const testResult = await context.pinion.exec('npm', ['run', 'compile'], { cwd })
+
+          assert.strictEqual(testResult, 0)
+        }
       })
     })
   }